requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
ccxt==4.3.11
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
dnspython==2.6.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
motor==3.4.0
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
pycares==4.4.0
pycparser==2.22
pymongo==4.7.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
wheel==0.43.0
yarl==1.9.4
